requirements.txt
Django==4.0.0
djangorestframework==3.12.4
pytz==2021.3
django-filter==22.1
djangorestframework-simplejwt
psycopg2-binary==2.8.6
requests==2.22.0
uwsgi==2.0.20